test(ci): fix flaky AlphabetRail timeout and consolidate redundant test files
CI runners oversubscribe 4 forks on a 4-core machine, starving jsdom env
setup. A trivial AlphabetRail render measured 5953ms wall-clock under
contention (vs 84ms isolated), exceeding vitest's default 5000ms
testTimeout — flaky timeout failure with no real bug.

- vite.config.ts: raise testTimeout/hookTimeout to 15s for CI contention
  headroom; cap poolOptions.forks.maxThreads to 3 on CI so the main
  process + v8 coverage instrumentation get headroom.
- Consolidate 3 FullscreenPlayerView test files (main + events + romanize)
  into 1 file with 3 tests. All shared the same mocks; 2 of 3 already
  needed jsdom. Saves 2 jsdom env setups.
- Remove redundant #113 createCdgFrameCoordinator describe block from
  use-cdg-sync.test.ts (2 tests). Superseded by the dedicated
  use-cdg-sync.coordinator.test.ts which tests the same behavior more
  thoroughly.

Coverage unchanged: Statements 83.05%, Branches 76.07%, Functions 75.55%,
Lines 83.7%. The 2 removed tests' code paths are fully covered by
.coordinator.test.ts.
