+++ src/hooks/use-cdg-sync.test.ts
@@ -1,10 +1,6 @@
 import { describe, expect, test, vi } from "vitest";
 import { ensureArrayBuffer } from "@/lib/cdg-protocol";
-import {
-  createCdgFrameCoordinator,
-  getCdgSyncBucket,
-  startCdgPositionSync,
-} from "./use-cdg-sync";
+import { getCdgSyncBucket, startCdgPositionSync } from "./use-cdg-sync";
 
 describe("ensureArrayBuffer", () => {
   test("returns ArrayBuffer as-is", () => {
@@ -224,132 +220,5 @@ describe("CDG frame IPC validates against current song before drawFrame", () => 
 
     expect(afterIpc).toContain("snapshot?.song_id");
     expect(afterIpc).toContain("transport_generation");
-  });
-});
-
-describe("#113: createCdgFrameCoordinator under slow IPC", () => {
-  function deferred<T>() {
-    let resolve!: (value: T) => void;
-    const promise = new Promise<T>((res) => {
-      resolve = res;
-    });
-    return { promise, resolve };
-  }
-
-  test("does not drop frame when a newer request was enqueued during slow IPC", async () => {
-    const first = deferred<ArrayBuffer>();
-    const second = deferred<ArrayBuffer>();
-    let n = 0;
-    const getCdgFrame = vi.fn(() => {
-      n += 1;
-      return n === 1 ? first.promise : second.promise;
-    });
-    const onFrame = vi.fn();
-    const onProbeResolved = vi.fn();
-    const onError = vi.fn();
-
-    const coordinator = createCdgFrameCoordinator({
-      getCdgFrame: getCdgFrame as never,
-      getCdgStatus: vi.fn().mockResolvedValue({
-        availability: "none",
-        songId: null,
-        transportGeneration: null,
-        packetCount: null,
-        errorCode: null,
-      }),
-      onFrame,
-      onProbeResolved,
-      onError,
-      isCurrent: () => true, // same song/generation throughout
-    });
-
-    // Enqueue request A (position 100).
-    coordinator.request({
-      songId: "song-a",
-      transportGeneration: 1,
-      positionMs: 100,
-      lastFrameVersion: 0,
-    });
-    expect(getCdgFrame).toHaveBeenCalledTimes(1);
-
-    // While A is in flight, enqueue request B (position 200).
-    coordinator.request({
-      songId: "song-a",
-      transportGeneration: 1,
-      positionMs: 200,
-      lastFrameVersion: 1,
-    });
-    expect(getCdgFrame).toHaveBeenCalledTimes(1);
-
-    // Now resolve A. The serial has not changed (only invalidate changes it),
-    // so A's result is processed.
-    first.resolve(new ArrayBuffer(0));
-    // Drain promise microtasks: then → finally → pump → second getCdgFrame.
-    for (let i = 0; i < 10; i++) {
-      await Promise.resolve();
-    }
-
-    expect(onProbeResolved).toHaveBeenCalledTimes(1);
-    expect(onProbeResolved).toHaveBeenCalledWith({
-      songId: "song-a",
-      transportGeneration: 1,
-      hasCdg: false,
-      hasFrame: false,
-      availability: "none",
-      errorCode: null,
-    });
-
-    expect(getCdgFrame).toHaveBeenCalledTimes(2);
-
-    second.resolve(new ArrayBuffer(0));
-    for (let i = 0; i < 10; i++) {
-      await Promise.resolve();
-    }
-    coordinator.invalidate();
-  });
-
-  test("drops frame when song changes (isCurrent returns false)", async () => {
-    const pending = deferred<ArrayBuffer>();
-    const getCdgFrame = vi.fn(() => pending.promise);
-    const onFrame = vi.fn();
-    const onProbeResolved = vi.fn();
-    const onError = vi.fn();
-
-    let currentSong = "song-a";
-    const coordinator = createCdgFrameCoordinator({
-      getCdgFrame: getCdgFrame as never,
-      getCdgStatus: vi.fn().mockResolvedValue({
-        availability: "none",
-        songId: null,
-        transportGeneration: null,
-        packetCount: null,
-        errorCode: null,
-      }),
-      onFrame,
-      onProbeResolved,
-      onError,
-      isCurrent: (req) => req.songId === currentSong,
-    });
-
-    // Enqueue request for song-a.
-    coordinator.request({
-      songId: "song-a",
-      transportGeneration: 1,
-      positionMs: 100,
-      lastFrameVersion: 0,
-    });
-
-    // Song changes while request is in flight.
-    currentSong = "song-b";
-
-    pending.resolve(new ArrayBuffer(0));
-    for (let i = 0; i < 10; i++) {
-      await Promise.resolve();
-    }
-
-    expect(onProbeResolved).not.toHaveBeenCalled();
-    expect(onError).not.toHaveBeenCalled();
-
-    coordinator.invalidate();
   });
 });
